+++ samples/Sample_Header.php
@@ -1,14 +1,10 @@
 <?php
-require_once __DIR__ . '/../src/PhpWord/Autoloader.php';
+require_once __DIR__ . '/../bootstrap.php';
 
-date_default_timezone_set('UTC');
-
-/**
- * Header file
- */
 use PhpOffice\PhpWord\Autoloader;
 use PhpOffice\PhpWord\Settings;
 
+date_default_timezone_set('UTC');
 error_reporting(E_ALL);
 define('CLI', (PHP_SAPI == 'cli') ? true : false);
 define('EOL', CLI ? PHP_EOL : '<br />');