+++ src/PhpWord/SimpleType/Jc.php
@@ -17,6 +17,8 @@
 
 namespace PhpOffice\PhpWord\SimpleType;
 
+use Zend\Validator\InArray;
+
 /**
  * Horizontal Alignment Type.
  *
@@ -36,22 +38,44 @@ final class Jc
     const THAI_DISTRIBUTE = 'thaiDistribute';
 
     /**
+     * @deprecated 0.13.0 Use `START` instead.
+     */
+    const LEFT = 'left';
+    /**
+     * @deprecated 0.13.0 Use `END` instead.
+     */
+    const RIGHT = 'right';
+    /**
+     * @deprecated 0.13.0 Use `BOTH` instead.
+     */
+    const JUSTIFY = 'justify';
+
+    /**
      * @since 0.13.0
      *
-     * @return string[]
+     * @return \Zend\Validator\InArray
      */
-    final public static function getAllowedValues()
-    {
-        return array(
-            self::START,
-            self::CENTER,
-            self::END,
-            self::MEDIUM_KASHIDA,
-            self::DISTRIBUTE,
-            self::NUM_TAB,
-            self::HIGH_KASHIDA,
-            self::LOW_KASHIDA,
-            self::THAI_DISTRIBUTE,
+    final public static function getValidator() {
+        // todo: consider caching validator instances.
+        return new InArray(
+            array (
+                'haystack' => array(
+                    self::START,
+                    self::CENTER,
+                    self::END,
+                    self::BOTH,
+                    self::MEDIUM_KASHIDA,
+                    self::DISTRIBUTE,
+                    self::NUM_TAB,
+                    self::HIGH_KASHIDA,
+                    self::LOW_KASHIDA,
+                    self::THAI_DISTRIBUTE,
+                    self::LEFT,
+                    self::RIGHT,
+                    self::JUSTIFY,
+                ),
+                'strict'   => InArray::COMPARE_STRICT,
+            )
         );
     }
 }